tuber: Client for the YouTube API

Get comments posted on YouTube videos, information on how many times a video has been liked, search for videos with particular content, and much more. You can also scrape captions from a few videos. To learn more about the YouTube API, see https://developers.google.com/youtube/v3/.

See also

Author

Maintainer: Gaurav Sood gsood07@gmail.com

Other contributors:

Kate Lyons k.lyons7@gmail.com [contributor]

John Muschelli muschellij2@gmail.com [contributor]