Search YouTube

Search for videos, channels and playlists. (By default, the function searches for videos.)
yt_search(term = NULL, max_results = 50, channel_id = NULL,
  channel_type = NULL, type = "video", event_type = NULL,
  location = NULL, location_radius = NULL, published_after = NULL,
  published_before = NULL, video_definition = "any",
  video_caption = "any", video_license = "any", video_syndicated = "any",
  video_type = "any", simplify = TRUE, get_all = TRUE,
  page_token = NULL, ...)

Arguments

term 	Character. Search term; required; no default

max_results 	Maximum number of items that should be returned. Integer. Optional. Can be between 0 and 50. Default is 50. Search results are constrained to a maximum of 500 videos if type is video and we have a value of channel_id.

channel_id 	Character. Only return search results from this channel; Optional.

channel_type 	Character. Optional. Takes one of two values: 'any', 'show'. Default is 'any'

type 	Character. Optional. Takes one of three values: 'video', 'channel', 'playlist'. Default is 'video'.

event_type 	Character. Optional. Takes one of three values: 'completed', 'live', 'upcoming'

location 	Character. Optional. Latitude and Longitude within parentheses, e.g. "(37.42307,-122.08427)"

location_radius 	Character. Optional. e.g. "1500m", "5km", "10000ft", "0.75mi"

published_after 	Character. Optional. RFC 339 Format. For instance, "1970-01-01T00:00:00Z"

published_before 	Character. Optional. RFC 339 Format. For instance, "1970-01-01T00:00:00Z"

video_definition 	Character. Optional. Takes one of three values: 'any' (return all videos; Default), 'high', 'standard'

video_caption 	Character. Optional. Takes one of three values: 'any' (return all videos; Default), 'closedCaption', 'none'. Type must be set to video.

video_license 	Character. Optional. Takes one of three values: 'any' (return all videos; Default), 'creativeCommon' (return videos with Creative Commons license), 'youtube' (return videos with standard YouTube license).

video_syndicated 	Character. Optional. Takes one of two values: 'any' (return all videos; Default), 'true' (return only syndicated videos)

video_type 	Character. Optional. Takes one of three values: 'any' (return all videos; Default), 'episode' (return episode of shows), 'movie' (return movies)

simplify 	Boolean. Return a data.frame if TRUE. Default is TRUE. If TRUE, it returns a list that carries additional information.

get_all 	get all results, iterating through all the results pages. Default is TRUE. Result is a data.frame. Optional.

page_token 	specific page in the result set that should be returned, optional

… 	Additional arguments passed to tuber_GET.

Value
data.frame with 16 elements: video_id, publishedAt, channelId, title, description, thumbnails.default.url, thumbnails.default.width, thumbnails.default.height, thumbnails.medium.url, thumbnails.medium.width, thumbnails.medium.height, thumbnails.high.url, thumbnails.high.width, thumbnails.high.height, channelTitle, liveBroadcastContent
References
https://developers.google.com/youtube/v3/docs/search/list
Examples

## Not run: ------------------------------------
# 
# # Set API token via yt_oauth() first
# 
# yt_search(term = "Barack Obama")
# yt_search(term = "Barack Obama", published_after = "2016-10-01T00:00:00Z")
# yt_search(term = "Barack Obama", published_before = "2016-09-01T00:00:00Z")
# yt_search(term = "Barack Obama", published_before = "2016-03-01T00:00:00Z", 
#                                published_after = "2016-02-01T00:00:00Z")
# yt_search(term = "Barack Obama", published_before = "2016-02-10T00:00:00Z", 
#                                published_after = "2016-01-01T00:00:00Z")
## ---------------------------------------------